# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple documentation on the gateway service which is a part of backend homework of WebDev course in spring 2023.",
        "title": "Web 2023 Backend HW - Gateway service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MohammadMahdi Gheidi",
            "email": "gheidimahdi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
        "/auth/req_dh_params": {
            "get": {
                "description": "after receiving p/g params from req_pq, communicate the keys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "req_dh_params"
                ],
                "summary": "handshake on diffie-hellman parameters between server/client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of message that will be sent to server",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "client's nonce",
                        "name": "nonce",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "server nonce that was sent to the client in the req_pg request",
                        "name": "server_nonce",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "the A value of the diffie-hellman",
                        "name": "a",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/req_pq": {
            "get": {
                "description": "this endpoint will request the p and g parameters of the diffie-hellman",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "req_pq"
                ],
                "summary": "request p/g params diffie-hellman",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of message that will be sent to server",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "client's nonce",
                        "name": "nonce",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/biz/get_users": {
            "get": {
                "description": "this endpoint will fetch the user with the id provided in the request or if no id is provided will return last 100 rows of db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get_users"
                ],
                "summary": "lets you do a query on db with secure functionality and no injection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of message that will be sent to server",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of user in the database",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "authenticated key that was calculated by both server and client",
                        "name": "auth_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/biz/get_users_with_sql_injection": {
            "get": {
                "description": "will not perform any security check on the input given by user and will execute the query without any checks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get_users_with_sql_injection"
                ],
                "summary": "enables you to perform a query on the db with injection feature",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of message that will be sent to server",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of user in the database or the dirty string which is injected",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "authenticated key that was calculated by both server and client",
                        "name": "auth_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}